requirements.txt
streamlit
groq
langchain
langchain-groq
chromadb
sentence-transformers
pymupdf
python-docx
python-dotenv